+++ Tests/Reqnroll.RuntimeTests/BoDi/ConcurrencyTests.cs
@@ -528,7 +528,9 @@ namespace Reqnroll.RuntimeTests.BoDi
                 var thread2 = new Thread(_ =>
                 {
                     Action act = () => container.Resolve<BlockingObject>();
-                    act.Should().ThrowExactly<ObjectContainerException>("Concurrent object resolution timeout (potential circular dependency).");
+                    act.Should().ThrowExactly<ObjectContainerException>("Concurrent object resolution timeout (potential circular dependency).")
+                       .And.Message.Should().Contain(nameof(ObjectContainer.DefaultConcurrentObjectResolutionTimeout))
+                       .And.Contain(nameof(ObjectContainer.ConcurrentObjectResolutionTimeout));
                 });
 
                 // start first thread and wait until ctor already in progress