+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Reduced MsBuild log output and consistent use of [Reqnroll] prefix (#381)
 * Update behavior of `ObjectContainer.IsRegistered()` to check base container for registrations, to match `Resolve()` behavior (#367)
 * Support loading plugin dependencies from .deps.json on .NET Framework and Visual Studio MSBuild (#408)
+* Support for setting `ObjectContainer.DefaultConcurrentObjectResolutionTimeout` even after creation of the container (#435)
 
 ## Bug fixes: